requirements.txt
# Core FastAPI and web framework
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.6

# Vector database
chromadb==0.4.18

# Document Loaders
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.0.2

# Only OpenAI - no other ML libraries
openai==1.6.1

# Data processing (minimal)
pandas==2.1.4
pydantic==2.5.2
pydantic-settings==2.1.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.1
httpx==0.25.2
aiofiles==23.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
black==23.11.0
starlette-prometheus==0.9.0

# Evaluation
rouge-score==0.1.2
nltk==3.8.1
datasets==2.15.0 # Para descargar datasets de Hugging Face 
scikit-learn==1.5.0
sentence-transformers==3.0.1
plotly==5.22.0
scipy==1.11.4  # Para bootstraping y estadísticas
pandas==2.1.4  # Para análisis de datos 